+++ src/lightinfoscrollarea.h
@@ -5,10 +5,8 @@
 #include <QVBoxLayout>
 #include <QWidget>
 
-#include "comm/arducor/arducorinfowidget.h"
 #include "comm/commhue.h"
 #include "comm/hue/hueinfowidget.h"
-#include "comm/nanoleaf/leaflightinfowidget.h"
 #include "cor/objects/page.h"
 #include "cor/widgets/topwidget.h"
 
@@ -30,42 +28,17 @@ public:
     void updateHues(std::vector<HueMetadata> lights);
 
     /*!
-     * \brief updateControllers update the controllers for nanoleafs to any
-     * hardware changes detected
-     *
-     * \param controllers nanoleaf controllers to use as the recent set.
-     */
-    void updateNanoLeafs(const std::vector<nano::LeafMetadata>& lights);
-
-    /// updates the arducor lights in the light info list widget
-    void updateAruCorLights(const std::vector<ArduCorMetadata>& lights);
-
-    /*!
      * \brief lookupCurrentLight looks up the current light based on what is currently selected
      * \return returns EProtocol::MAX if nothing is selected, otherwise, returns the current light
      */
-    std::pair<EProtocolType, QString> lookupCurrentLight();
+    QString lookupCurrentLight();
 
     /// getter for current key
-    QString key() const noexcept {
-        switch (mCurrentProtocol) {
-            case EProtocolType::hue:
-                return mLastHueKey;
-            case EProtocolType::arduCor:
-                return mLastArduCorKey;
-            case EProtocolType::nanoleaf:
-                return mLastNanoleafKey;
-            case EProtocolType::MAX:
-                return {};
-        }
-        return {};
-    }
+    QString key() const noexcept { return mLastHueKey; }
 
     /// resize programmatically
     void resize();
 
-    /// change the protocol currently displayed
-    void changeProtocol(EProtocolType);
 signals:
 
     /// emits the key and whether or not a light is selected when a light is clicked
@@ -86,23 +59,8 @@ private:
     /// widgets for hue displayed in scroll area
     std::vector<hue::HueInfoWidget*> mHueWidgets;
 
-    /// widgets for nanoleaf displayed in scroll area
-    std::vector<nano::LeafLightInfoWidget*> mNanoleafWidgets;
-
-    /// widgets for ArduCor displayed in scroll area
-    std::vector<ArduCorInfoWidget*> mArduCorWidgets;
-
     /// stores the last hue key clicked
     QString mLastHueKey;
-
-    /// stores the last arducor key clicked
-    QString mLastArduCorKey;
-
-    /// stores the last nanoleaf key clicked
-    QString mLastNanoleafKey;
-
-    /// stores the current protocol shown
-    EProtocolType mCurrentProtocol;
 };
 
 #endif // LIGHTINFOSCROLLAREA_H